Move key generation classes into own modulew, as this is intended to be separate to the main application
