+++ nexus-app/pom.xml
@@ -15,6 +15,8 @@
 
     <dependencies>
 
+        <!-- Nexus modules -->
+
         <dependency>
             <groupId>com.github.nexus</groupId>
             <artifactId>argon2</artifactId>
@@ -30,8 +32,15 @@
             <groupId>com.github.nexus</groupId>
             <artifactId>nexus-encryption</artifactId>
         </dependency>
-        
-        
+
+        <dependency>
+            <groupId>com.github.nexus</groupId>
+            <artifactId>nexus-keygen</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+
+
+
         <dependency>
             <groupId>org.glassfish.jersey.core</groupId>
             <artifactId>jersey-server</artifactId>
