+++ nexus-encryption-kalium/pom.xml
@@ -1,15 +1,20 @@
 <?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
     <modelVersion>4.0.0</modelVersion>
+
     <parent>
         <groupId>com.github.nexus</groupId>
         <artifactId>nexus</artifactId>
         <version>1.0-SNAPSHOT</version>
     </parent>
+
     <artifactId>nexus-encryption-kalium</artifactId>
     <packaging>jar</packaging>
-    
-    
+
+
     <dependencyManagement>
         <dependencies>
             <dependency>
@@ -17,23 +22,23 @@
                 <artifactId>kalium</artifactId>
                 <version>0.8.0</version>
             </dependency>
-    
-        
+
+
         </dependencies>
     </dependencyManagement>
-    
+
     <dependencies>
         <dependency>
             <groupId>com.github.nexus</groupId>
             <artifactId>nexus-encryption</artifactId>
         </dependency>
-        
+
         <dependency>
             <groupId>org.abstractj.kalium</groupId>
             <artifactId>kalium</artifactId>
         </dependency>
-    
-        
+
+
     </dependencies>
     <name>nexus-encryption-kalium</name>
 </project>
