+++ nexus-keygen/pom.xml
@@ -1,0 +1,55 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>nexus-keygen</artifactId>
+
+    <name>nexus-keygen</name>
+    <packaging>jar</packaging>
+
+    <parent>
+        <artifactId>nexus</artifactId>
+        <groupId>com.github.nexus</groupId>
+        <version>1.0-SNAPSHOT</version>
+    </parent>
+
+    <dependencyManagement>
+
+        <dependencies>
+
+        </dependencies>
+
+    </dependencyManagement>
+
+    <dependencies>
+
+        <dependency>
+            <groupId>com.github.nexus</groupId>
+            <artifactId>argon2</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>org.glassfish</groupId>
+            <artifactId>javax.json</artifactId>
+        </dependency>
+
+        <dependency>
+            <groupId>com.github.nexus</groupId>
+            <artifactId>nexus-configurator</artifactId>
+            <version>1.0-SNAPSHOT</version>
+        </dependency>
+
+        <dependency>
+            <groupId>com.github.nexus</groupId>
+            <artifactId>nexus-encryption</artifactId>
+        </dependency>
+
+    </dependencies>
+
+
+
+</project>
