+++ nexus-keygen/src/main/java/com/github/nexus/keygen/KeyGenerator.java
@@ -1,4 +1,4 @@
-package com.github.nexus.key;
+package com.github.nexus.keygen;
 
 import com.github.nexus.nacl.KeyPair;
